requirements.txt
numpy
opencv-python
ortools
pillow>=5.2.0
requests
progress
pandas
PyYAML
requests
matplotlib
numba
scipy
scikit-learn
cached-property
click
qrcode
scikit-image
watchdog